Refactoring (#177)
