+++ internal/provider/resource_integer.go
@@ -1,4 +1,4 @@
-package integer
+package provider
 
 import (
 	"context"
@@ -13,15 +13,11 @@ import (
 	"github.com/terraform-providers/terraform-provider-random/internal/random"
 )
 
-func NewResourceType() *resourceType {
-	return &resourceType{}
-}
+var _ tfsdk.ResourceType = (*integerResourceType)(nil)
 
-var _ tfsdk.ResourceType = (*resourceType)(nil)
+type integerResourceType struct{}
 
-type resourceType struct{}
-
-func (r *resourceType) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
+func (r *integerResourceType) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
 	return tfsdk.Schema{
 		Description: "The resource `random_integer` generates random values from a given range, described " +
 			"by the `min` and `max` attributes of a given resource.\n" +
@@ -71,19 +67,19 @@ func (r *resourceType) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostic
 	}, nil
 }
 
-func (r *resourceType) NewResource(_ context.Context, _ tfsdk.Provider) (tfsdk.Resource, diag.Diagnostics) {
-	return &resource{}, nil
+func (r *integerResourceType) NewResource(_ context.Context, _ tfsdk.Provider) (tfsdk.Resource, diag.Diagnostics) {
+	return &integerResource{}, nil
 }
 
 var (
-	_ tfsdk.Resource                = (*resource)(nil)
-	_ tfsdk.ResourceWithImportState = (*resource)(nil)
+	_ tfsdk.Resource                = (*integerResource)(nil)
+	_ tfsdk.ResourceWithImportState = (*integerResource)(nil)
 )
 
-type resource struct{}
+type integerResource struct{}
 
-func (r *resource) Create(ctx context.Context, req tfsdk.CreateResourceRequest, resp *tfsdk.CreateResourceResponse) {
-	var plan modelV0
+func (r *integerResource) Create(ctx context.Context, req tfsdk.CreateResourceRequest, resp *tfsdk.CreateResourceResponse) {
+	var plan integerModelV0
 
 	diags := req.Plan.Get(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
@@ -106,7 +102,7 @@ func (r *resource) Create(ctx context.Context, req tfsdk.CreateResourceRequest, 
 	rand := random.NewRand(seed)
 	number := rand.Intn((max+1)-min) + min
 
-	u := &modelV0{
+	u := &integerModelV0{
 		ID:      types.String{Value: strconv.Itoa(number)},
 		Keepers: plan.Keepers,
 		Min:     types.Int64{Value: int64(min)},
@@ -128,20 +124,20 @@ func (r *resource) Create(ctx context.Context, req tfsdk.CreateResourceRequest, 
 }
 
 // Read does not need to perform any operations as the state in ReadResourceResponse is already populated.
-func (r *resource) Read(ctx context.Context, req tfsdk.ReadResourceRequest, resp *tfsdk.ReadResourceResponse) {
+func (r *integerResource) Read(ctx context.Context, req tfsdk.ReadResourceRequest, resp *tfsdk.ReadResourceResponse) {
 }
 
 // Update is intentionally left blank as all required and optional attributes force replacement of the resource
 // through the RequiresReplace AttributePlanModifier.
-func (r *resource) Update(ctx context.Context, req tfsdk.UpdateResourceRequest, resp *tfsdk.UpdateResourceResponse) {
+func (r *integerResource) Update(ctx context.Context, req tfsdk.UpdateResourceRequest, resp *tfsdk.UpdateResourceResponse) {
 }
 
 // Delete does not need to explicitly call resp.State.RemoveResource() as this is automatically handled by the
 // [framework](https://github.com/hashicorp/terraform-plugin-framework/pull/301).
-func (r *resource) Delete(ctx context.Context, req tfsdk.DeleteResourceRequest, resp *tfsdk.DeleteResourceResponse) {
+func (r *integerResource) Delete(ctx context.Context, req tfsdk.DeleteResourceRequest, resp *tfsdk.DeleteResourceResponse) {
 }
 
-func (r *resource) ImportState(ctx context.Context, req tfsdk.ImportResourceStateRequest, resp *tfsdk.ImportResourceStateResponse) {
+func (r *integerResource) ImportState(ctx context.Context, req tfsdk.ImportResourceStateRequest, resp *tfsdk.ImportResourceStateResponse) {
 	parts := strings.Split(req.ID, ",")
 	if len(parts) != 3 && len(parts) != 4 {
 		resp.Diagnostics.AddError(
@@ -181,7 +177,7 @@ func (r *resource) ImportState(ctx context.Context, req tfsdk.ImportResourceStat
 		return
 	}
 
-	var state modelV0
+	var state integerModelV0
 
 	state.ID.Value = parts[0]
 	state.Keepers.ElemType = types.StringType
@@ -200,7 +196,7 @@ func (r *resource) ImportState(ctx context.Context, req tfsdk.ImportResourceStat
 	}
 }
 
-type modelV0 struct {
+type integerModelV0 struct {
 	ID      types.String `tfsdk:"id"`
 	Keepers types.Map    `tfsdk:"keepers"`
 	Min     types.Int64  `tfsdk:"min"`
